release 0.4.7
- Almost complete colored solver messages (in 'report' etc.).
- Support for reading and encoding XNF (CNF with OR and XOR clauses).
- New makefile goals: test-two-ways, test-all-pairs, test-all-triples.
- Binary search in two-way combinatorial tester 'gencombi'.
- Code clean-up: sorted statistics fields, moved some internal code up.
- Using stable radix sort in 'reduce' allowed to remove clause ids.
- Radix sorting (to speed-up sorting of 'analyzed' stack).

+++ tatch.sh
@@ -56,19 +56,23 @@ run () {
   then
     cnf="$2"
     case "$cnf" in
-      *.cnf);;
+      *.cnf) xnf=no;;
+      *.xnf) xnf=yes;;
       *) die "expected CNF as second argument in '$expected $command'";;
     esac
-    proofpath=cnfs/`basename $cnf .cnf`.proof
-    rm -f $proofpath 2>/dev/null || die "failed to 'rm -f $proofpath'"
-    options=""
-    case $proofsmod3 in
-      0) proofsmod3=1; ;;
-      1) proofsmod3=2; proof="$proofpath";;
-      2) proofsmod3=0; options="-a"; proof="$proofpath";;
-    esac
-    [ "$options" ] && command="$command $options"
-    [ "$proof" ] && command="$command $proof"
+    if [ $xnf = no ]
+    then
+      proofpath=cnfs/`basename $cnf .cnf`.proof
+      rm -f $proofpath 2>/dev/null || die "failed to 'rm -f $proofpath'"
+      options=""
+      case $proofsmod3 in
+	0) proofsmod3=1; ;;
+	1) proofsmod3=2; proof="$proofpath";;
+	2) proofsmod3=0; options="-a"; proof="$proofpath";;
+      esac
+      [ "$options" ] && command="$command $options"
+      [ "$proof" ] && command="$command $proof"
+    fi
   fi
   printf "$command # expected '$expected'"
   $command 1>/dev/null 2>/dev/null
@@ -118,6 +122,12 @@ msg "now solving CNF files"
 run 10 ./satch cnfs/true.cnf
 run 20 ./satch cnfs/false.cnf
 
+run 10 ./satch xnfs/true.xnf
+run 20 ./satch xnfs/false.xnf
+
+run 10 ./satch cnfs/trivial.cnf
+run 20 ./satch xnfs/inconsistent.xnf
+
 run 10 ./satch cnfs/unit1.cnf
 run 10 ./satch cnfs/unit2.cnf
 run 10 ./satch cnfs/unit3.cnf
@@ -125,9 +135,26 @@ run 10 ./satch cnfs/unit4.cnf
 run 20 ./satch cnfs/unit5.cnf
 run 20 ./satch cnfs/unit6.cnf
 
+run 10 ./satch xnfs/unit1.xnf
+run 10 ./satch xnfs/unit2.xnf
+run 20 ./satch xnfs/unit3.xnf
+
 run 10 ./satch cnfs/unit7.cnf
 run 20 ./satch cnfs/unit8.cnf
 run 20 ./satch cnfs/unit9.cnf
+
+run 10 ./satch xnfs/xor2.xnf
+run 10 ./satch xnfs/xor3.xnf
+run 10 ./satch xnfs/xor4.xnf
+run 10 ./satch xnfs/xor5.xnf
+run 10 ./satch xnfs/xor6.xnf
+run 10 ./satch xnfs/xor7.xnf
+run 10 ./satch xnfs/xor8.xnf
+run 10 ./satch xnfs/xor9.xnf
+run 10 ./satch xnfs/xor10.xnf
+run 10 ./satch xnfs/xor11.xnf
+run 10 ./satch xnfs/xor12.xnf
+run 10 ./satch xnfs/xor24.xnf
 
 run 20 ./satch cnfs/full2.cnf
 run 20 ./satch cnfs/full3.cnf
@@ -174,6 +201,8 @@ run 10 ./satch cnfs/prime1369.cnf
 run 10 ./satch cnfs/prime1681.cnf
 run 10 ./satch cnfs/prime1849.cnf
 run 10 ./satch cnfs/prime2209.cnf
+
+run 10 ./satch cnfs/regr1.cnf
 
 [ $learning = no ] && \
 run 20 ./satch cnfs/prime65537.cnf
